refac: datadetailpage convention 맞게 수정
