+++ test-web/src/components/DataDetailPage/AcceptModal.js
@@ -4,6 +4,8 @@ import { Backdrop, Box, Modal, Fade, Button, Typography } from '@mui/material';
 import { FaRegCheckCircle } from 'react-icons/fa';
 import StateChangedModal from './StateChangedModal';
 import updateDataStatus from '../../API/updateDataStatus';
+import style from './style/acceptmodalstyle';
+
 
 const navy = '#0F3659';
 
@@ -136,16 +138,4 @@ export const AcceptModal = ({ meatId, setConfirmVal, confirmVal }) => {
   );
 };
 export default AcceptModal;
-
-const style = {
-  position: 'absolute',
-  top: '50%',
-  left: '50%',
-  transform: 'translate(-50%, -50%)',
-  width: 500,
-  bgcolor: 'background.paper',
-  boxShadow: 24,
-  p: 4,
-  borderRadius: '10px',
-};
