+++ test-web/src/components/DataDetailPage/DataConfirmView.js
@@ -37,6 +37,7 @@ import RejectModal from './RejectModal';
 import { FaRegCheckCircle, FaRegTimesCircle } from 'react-icons/fa';
 // mui
 import { IconButton, TextField, Autocomplete } from '@mui/material';
+import {style, divStyle} from './style/dataviewstyle';
 
 // import tables
 // import RawTable from './tablesComps/rawTable';
@@ -411,137 +412,3 @@ const apiField = [
   'statusType',
   'traceNum',
 ];
-
-const style = {
-  singleDataWrapper: {
-    marginTop: '40px',
-    display: 'flex',
-    justifyContent: 'space-between',
-  },
-  confirmEditBtnWrapper: {
-    paddingTop: '0px',
-    width: '100%',
-    display: 'flex',
-    justifyContent: 'end',
-    marginTop: 'auto',
-    paddingRight: '20px',
-    borderBottomLeftRadius: '10px',
-    borderBottomRightRadius: '10px',
-    minWidth: '1140px',
-  },
-  acceptBtn: {
-    backgroundColor: '#00e676',
-    color: 'white',
-    fontSize: '15px',
-    borderRadius: '5px',
-    width: '80px',
-    height: '35px',
-  },
-  rejectBtn: {
-    backgroundColor: '#e53935',
-    color: 'white',
-    fontSize: '15px',
-    borderRadius: '5px',
-    width: '80px',
-    height: '35px',
-    marginLeft: '20px',
-  },
-  editBtnWrapper: {
-    padding: '0px',
-    margin: '10px 0px 0px',
-    paddingRight: '10px',
-    width: '100%',
-    display: 'flex',
-    justifyContent: 'end',
-    minWidth: '1140px',
-  },
-  dataFieldColumn: {
-    backgroundColor: '#9e9e9e',
-    height: '33px',
-    borderRight: '1px solid rgb(174, 168, 168)',
-    borderBottom: '1px solid #fafafa',
-    padding: '4px 5px',
-  },
-  dataExpColumn: {
-    backgroundColor: '#757575',
-    height: '33px',
-    borderRight: '1px solid rgb(174, 168, 168)',
-    borderBottom: '1px solid #fafafa',
-    padding: '4px 5px',
-    color: 'white',
-  },
-  dataFieldContainer: {
-    backgroundColor: '#eeeeee',
-    height: '100%',
-    borderRight: '1px solid rgb(174, 168, 168)',
-    borderBottom: '1px solid #fafafa',
-    padding: '4px 5px',
-  },
-  dataContainer: {
-    height: '33px',
-    borderBottom: '0.8px solid #e0e0e0',
-    width: '60px',
-    borderRight: '0.8px solid #e0e0e0',
-    padding: '4px 5px',
-  },
-  completeBtn: {
-    border: `1px solid ${navy}`,
-    color: `${navy}`,
-    borderRadius: '5px',
-    width: '60px',
-    height: '35px',
-  },
-  editBtn: {
-    backgroundColor: `${navy}`,
-    border: 'none',
-    color: 'white',
-    borderRadius: '5px',
-    width: '60px',
-    height: '35px',
-  },
-};
-const divStyle = {
-  currDiv: {
-    height: 'fit-content',
-    width: 'fit-content',
-    padding: '10px',
-    borderRadius: '5px',
-    color: navy,
-  },
-  notCurrDiv: {
-    height: '100%',
-    width: 'fit-content',
-    borderRadius: '5px',
-    padding: '10px',
-    color: '#b0bec5',
-  },
-  loadingBackground: {
-    position: 'absolute',
-    width: '100vw',
-    height: '100vh',
-    top: 0,
-    left: 0,
-    backgroundColor: '#ffffffb7',
-    zIndex: 999,
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  loadingText: {
-    fontSize: '25px',
-    textAlign: 'center',
-  },
-  errorContainer: {
-    height: '65vh',
-    minHeight: '500px',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    height: '100%',
-  },
-  errorText: {
-    fontSize: '16px',
-    textAlign: 'center',
-  },
-};
